QuickAdminPanel automatic commit

+++ database/migrations/2020_11_18_000020_add_relationship_fields_to_teams_table.php
@@ -9,7 +9,7 @@ class AddRelationshipFieldsToTeamsTable extends Migration
     public function up()
     {
         Schema::table('teams', function (Blueprint $table) {
-            $table->unsignedInteger('owner_id');
+            $table->unsignedBigInteger('owner_id');
             $table->foreign('owner_id', 'owner_fk_2472571')->references('id')->on('users');
         });
     }
