+++ database/migrations/2020_11_18_000021_add_relationship_fields_to_garages_table.php
@@ -9,9 +9,9 @@ class AddRelationshipFieldsToGaragesTable extends Migration
     public function up()
     {
         Schema::table('garages', function (Blueprint $table) {
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id', 'user_fk_2578922')->references('id')->on('users');
-            $table->unsignedInteger('team_id')->nullable();
+            $table->unsignedBigInteger('team_id')->nullable();
             $table->foreign('team_id', 'team_fk_2587510')->references('id')->on('teams');
         });
     }
